Percents of the pool
My first solidity selfwritten contract. Yay!

It's allows to start a pools, imitating gamble sessions. 
With every new participant in the pool the persantages of winning of every participant changes, depending on he's bet. 

There is no Roll function yet, but it will be. 
Later there will be addresses, not only dummy names. 

If somebody wants to make a Roll function - you're welcome!