Imitation of a gamble pool game
My first solidity smart contract. Yay!

It is imitating a gambling session game, where you can start a pool and "participate" in it. 
With every new Participant percents of winning changes of every new participant, depending on their bet and total bank of a session. 

It is sandbox for me, so no doubt it is raw. 
Possibly, there will be more confirmations and restrictions for correct usability. 
Also there should be the Roll function. If you are want to participate in here - you are welcome!